Replace ICBINBSN hot example with RetainScene version
