+++ crates/bevy_bsn/Cargo.toml
@@ -11,12 +11,9 @@ bevy = { workspace = true }
 thiserror = { workspace = true }
 variadics_please = { workspace = true }
 
-[dev-dependencies]
-bevy_i-cant-believe-its-not-bsn.workspace = true
-
 [lints]
 workspace = true
 
 [[example]]
-name = "bsn_icbinbsn_hot"
+name = "bsn_asset_hot"
 required-features = ["bevy/file_watcher"]
